Added enum ImageFormat::NumberFormat which is {FLOATING_POINT_FORMAT, INTEGER_FORMAT, NORMALIZED_FIXED_POINT_FORMAT, OTHER}.  To eventually replace the floatingPoint boolean. Modified the constructor to match, and set its value for all image formats specified in the .cpp.
 Added ImageFormat::isIntegerFormat()

 Added ImageFormat::channelMask, which returns a Color4 that masks off unused components in the format, given in RGBA. 
 For example, the mask for R32F is (1,0,0,0), for A32F is (0,0,0,1), for RGB32F is (1,1,1,0). 
 (Note that luminance is interpreted as using only the R channel). Used by GLCaps::supportsTextureDrawBuffer().

 Added GLCaps::supportsTextureDrawBuffer(const ImageFormat* fmt). Extended RenderDevice::init to write out the 
 result of this function for each format alongside the supportsTexture and supportsRenderBuffer functions.

 BugFix: G3D::ImageFormat now correctly prints RGB9E5F as "RGB9E5F", instead of "RGB9E10F".

git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1336 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/GLG3D.lib/source/GFont.cpp
@@ -213,14 +213,14 @@ Vector2 GFont::drawString
             }
 
             float xx = x - sx;
-            glTexCoord2f(col * charWidth, row * charHeight + 1);
+            glTexCoord2f(float(col * charWidth), float(row * charHeight + 1));
             glVertex2f(xx,     y + sy);
 
-            glTexCoord2f(col * charWidth, (row + 1) * charHeight - 2);
+            glTexCoord2f(float(col * charWidth), float((row + 1) * charHeight - 2));
             glVertex2f(xx,     y + h - sy); 
 
             xx += w;
-            glTexCoord2f((col + 1) * charWidth - 1, (row + 1) * charHeight - 2);
+            glTexCoord2f(float((col + 1) * charWidth - 1), float((row + 1) * charHeight - 2));
             glVertex2f(xx, y + h - sy); 
 
             glTexCoord2f(float((col + 1) * charWidth - 1), float(row * charHeight + 1));
@@ -283,8 +283,8 @@ Vector2 GFont::computePackedArray(
             array[count].y = y + sy;
 
             ++count;
-            array[count].x = col * charWidth;
-            array[count].y = (row + 1) * charHeight - 2;
+            array[count].x = float(col * charWidth);
+            array[count].y = float((row + 1) * charHeight - 2);
             
             ++count;
             array[count].x = xx;
@@ -292,16 +292,16 @@ Vector2 GFont::computePackedArray(
 
             xx += w;
             ++count;
-            array[count].x = (col + 1) * charWidth - 1;
-            array[count].y = (row + 1) * charHeight - 2;
+            array[count].x = float((col + 1) * charWidth - 1);
+            array[count].y = float((row + 1) * charHeight - 2);
             
             ++count;
             array[count].x = xx;
             array[count].y = y + h - sy;
     
             ++count;
-            array[count].x = (col + 1) * charWidth - 1;
-            array[count].y = row * charHeight + 1;
+            array[count].x = float((col + 1) * charWidth - 1);
+            array[count].y = float(row * charHeight + 1);
 
             ++count;
             array[count].x = xx;
